+++ framework/console/console-gwt2/src/main/java/org/jmonitoring/console/gwt/client/methodcall/distribution/MethodCallDistributionPlace.java
@@ -16,7 +16,7 @@ public class MethodCallDistributionPlace extends Place implements ActivityAware
 
     int position;
 
-    int gapDuration;
+    long gapDuration;
 
     @Prefix("methodcallstat")
     public static class MethodCallStatTokenizer extends AbstractPlaceTokenizer implements
